Split tests module from command_interpreter.rs into dedicated file
